requirements.txt
Flask==1.0.0
Pillow==6.1.0
removebg==0.3
Werkzeug==0.14.1
